requirements.txt
torch
torchvision
streamlit
opencv-python
matplotlib
ultralytics
pillow
numpy
scikit-learn
seaborn
tqdm